requirements.txt
numpy
django-eventstream
channels==3.0.5
django-cors-headers
